move helpers in to relationhelper.py
